Small code cleaning
