+++ bow/bowgetters.go
@@ -6,15 +6,6 @@ import (
 	"github.com/apache/arrow/go/arrow/array"
 )
 
-func (b *bow) GetColNameIndex(s string) int {
-	for i, f := range b.Schema().Fields() {
-		if f.Name == s {
-			return i
-		}
-	}
-	return -1
-}
-
 func (b *bow) GetRow(rowIndex int) map[string]interface{} {
 	m := map[string]interface{}{}
 	for colIndex := 0; colIndex < b.NumCols(); colIndex++ {
@@ -278,4 +269,14 @@ func (b *bow) GetIndex(colName string) (int, error) {
 		}
 	}
 	return 0, fmt.Errorf("no column '%s'", colName)
+}
+
+// Unused and similar to function GetIndex above
+func (b *bow) GetColNameIndex(s string) int {
+	for i, f := range b.Schema().Fields() {
+		if f.Name == s {
+			return i
+		}
+	}
+	return -1
 }
