+++ bow/rolling/aggregation/XXXbenchmarks_test.go
@@ -2,11 +2,10 @@ package aggregation
 
 import (
 	"fmt"
-	"math/rand"
-	"testing"
-
 	"github.com/metronlab/bow/bow"
 	"github.com/metronlab/bow/bow/rolling"
+	"math/rand"
+	"testing"
 )
 
 // BenchSize of 1e8 triggers out of memory on a 16Go mem computer
