+++ bow/rolling/aggregation/count.go
@@ -7,7 +7,7 @@ import (
 
 func Count(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, false, bow.Int64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			var count int64
 			for i := 0; i < w.Bow.NumRows(); i++ {
 				v := w.Bow.GetValue(col, i)
