+++ bow/rolling/aggregation/difference.go
@@ -10,7 +10,7 @@ func Difference(col string) rolling.ColumnAggregation {
 	last := Last(col).Func()
 
 	return rolling.NewColumnAggregation(col, true, bow.Float64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			if w.Bow.NumRows() == 0 {
 				return nil, nil
 			}
