+++ bow/rolling/aggregation/difference_test.go
@@ -1,10 +1,9 @@
 package aggregation
 
 import (
-	"testing"
-
 	"github.com/metronlab/bow/bow"
 	"github.com/stretchr/testify/assert"
+	"testing"
 )
 
 func TestDifference(t *testing.T) {
