+++ bow/rolling/aggregation/integral.go
@@ -7,7 +7,7 @@ import (
 
 func IntegralTrapezoid(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, true, bow.Float64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			if w.Bow.NumRows() == 0 {
 				return nil, nil
 			}
@@ -39,7 +39,7 @@ func IntegralTrapezoid(col string) rolling.ColumnAggregation {
 
 func IntegralStep(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, false, bow.Float64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			if w.Bow.NumRows() == 0 {
 				return nil, nil
 			}
