+++ bow/rolling/aggregation/integral_test.go
@@ -1,11 +1,10 @@
 package aggregation
 
 import (
-	"testing"
-
 	"github.com/metronlab/bow/bow"
 	"github.com/metronlab/bow/bow/rolling/aggregation/transform"
 	"github.com/stretchr/testify/assert"
+	"testing"
 )
 
 func TestIntegralStep(t *testing.T) {
