+++ bow/rolling/aggregation/minmax.go
@@ -7,7 +7,7 @@ import (
 
 func Min(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, false, bow.Float64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			if w.Bow.NumRows() == 0 {
 				return nil, nil
 			}
@@ -32,7 +32,7 @@ func Min(col string) rolling.ColumnAggregation {
 
 func Max(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, false, bow.Float64,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			if w.Bow.NumRows() == 0 {
 				return nil, nil
 			}
