+++ bow/rolling/aggregation/windowstart.go
@@ -7,7 +7,7 @@ import (
 
 func WindowStart(col string) rolling.ColumnAggregation {
 	return rolling.NewColumnAggregation(col, false, bow.IteratorDependent,
-		func(col int, w bow.Window) (interface{}, error) {
+		func(col int, w rolling.Window) (interface{}, error) {
 			return w.Start, nil
 		})
 }
