+++ bow/rolling/fill/linear.go
@@ -7,7 +7,7 @@ import (
 
 func Linear(colName string) rolling.ColumnInterpolation {
 	return rolling.NewColumnInterpolation(colName, []bow.Type{bow.Int64, bow.Float64},
-		func(inputCol int, w bow.Window, full bow.Bow) (interface{}, error) {
+		func(inputCol int, w rolling.Window, full bow.Bow) (interface{}, error) {
 			t0, v0, prevIndex := full.GetPreviousFloat64s(w.IntervalColumnIndex, inputCol, w.FirstIndex-1)
 			if prevIndex == -1 {
 				return nil, nil
