+++ bow/rolling/fill_test.go
@@ -2,19 +2,18 @@ package rolling
 
 import (
 	"fmt"
-	"testing"
-
 	"github.com/metronlab/bow/bow"
 	"github.com/stretchr/testify/assert"
+	"testing"
 )
 
 func TestIntervalRolling_Fill(t *testing.T) {
 	timeInterp := NewColumnInterpolation(timeCol, []bow.Type{bow.Int64},
-		func(colIndex int, w bow.Window, full bow.Bow) (interface{}, error) {
+		func(colIndex int, w Window, full bow.Bow) (interface{}, error) {
 			return w.Start, nil
 		})
 	valueInterp := NewColumnInterpolation(valueCol, []bow.Type{bow.Int64, bow.Float64},
-		func(colIndex int, w bow.Window, full bow.Bow) (interface{}, error) {
+		func(colIndex int, w Window, full bow.Bow) (interface{}, error) {
 			return 9.9, nil
 		})
 
@@ -25,7 +24,7 @@ func TestIntervalRolling_Fill(t *testing.T) {
 		})
 		r, _ := IntervalRolling(b, timeCol, 2, Options{})
 		interp := NewColumnInterpolation(valueCol, []bow.Type{bow.Int64, bow.Bool},
-			func(colIndex int, w bow.Window, full bow.Bow) (interface{}, error) {
+			func(colIndex int, w Window, full bow.Bow) (interface{}, error) {
 				return true, nil
 			})
 		_, err := r.
