+++ bow/series.go
@@ -3,7 +3,6 @@ package bow
 import (
 	"errors"
 	"fmt"
-
 	"github.com/apache/arrow/go/arrow"
 	"github.com/apache/arrow/go/arrow/array"
 	"github.com/apache/arrow/go/arrow/memory"
